model/requirements.txt
numpy>=2.3.4
pydantic-settings>=2.11.0
python-dotenv>=1.2.1
scikit-learn>=1.7.2
sentence-transformers>=5.1.2
spacy>=3.8.7
requests>=2.31.0
pillow>=10.0.0
opencv-python>=4.8.0
ultralytics>=8.0.0
labelimg>=1.8.6
tqdm>=4.65.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
ultralytics>=8.3.222
pandas>=2.3.3
torchvision>=0.24.0
openai>=1.54.0
asyncio>=4.0.0
ddgs>=9.9.1
